# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EdgeIQ Market Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SignalType = (typeof SignalType)[keyof typeof SignalType];

export const SignalType = {
  insider: "insider",
  options: "options",
  sentiment: "sentiment",
} as const;

export interface Signal {
  id: number;
  ticker: string;
  type: SignalType;
  action: string;
  description: string;
  convictionScore: number;
  winRate: number;
  valueUsd?: number;
  reportedAt: string;
  filerName?: string;
  expiryDate?: string;
  strikePrice?: number;
  optionType?: string;
  sentiment?: string;
  source?: string;
}

export interface MarketQuote {
  ticker: string;
  price: number;
  change: number;
  changePercent: number;
  volume: number;
  marketCap?: string;
  pe?: number;
  high52w?: number;
  low52w?: number;
}

export type MarketStatsTopSectorsItem = {
  name: string;
  score: number;
};

export interface MarketStats {
  overallSentiment: string;
  fearGreedIndex: number;
  insiderBuyRatio: number;
  optionsFlowBias: string;
  totalSignalsToday: number;
  bullishSignals: number;
  bearishSignals: number;
  topSectors: MarketStatsTopSectorsItem[];
}

export interface WatchlistItem {
  id: number;
  ticker: string;
  addedAt: string;
  latestSignalScore?: number;
  alertsEnabled: boolean;
}

export interface AddWatchlistRequest {
  ticker: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type GetSignalsParams = {
  type?: GetSignalsType;
  ticker?: string;
  limit?: number;
};

export type GetSignalsType =
  (typeof GetSignalsType)[keyof typeof GetSignalsType];

export const GetSignalsType = {
  insider: "insider",
  options: "options",
  sentiment: "sentiment",
  all: "all",
} as const;