# lib/api-zod/src/generated/types/marketStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EdgeIQ Market Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketStatsTopSectorsItem } from "./marketStatsTopSectorsItem";

export interface MarketStats {
  overallSentiment: string;
  fearGreedIndex: number;
  insiderBuyRatio: number;
  optionsFlowBias: string;
  totalSignalsToday: number;
  bullishSignals: number;
  bearishSignals: number;
  topSectors: MarketStatsTopSectorsItem[];
}