# lib/api-zod/src/generated/types/signal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EdgeIQ Market Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { SignalType } from "./signalType";

export interface Signal {
  id: number;
  ticker: string;
  type: SignalType;
  action: string;
  description: string;
  convictionScore: number;
  winRate: number;
  valueUsd?: number;
  reportedAt: string;
  filerName?: string;
  expiryDate?: string;
  strikePrice?: number;
  optionType?: string;
  sentiment?: string;
  source?: string;
}