# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EdgeIQ Market Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistItem {
  id: number;
  ticker: string;
  addedAt: string;
  latestSignalScore?: number;
  alertsEnabled: boolean;
}